encouraging visitors to star projects they like
added stars for projects which are on github 
 * encourages visitors to "vote" for the projects they know and like by clicking on the starts button, and also meet other stargazers with similar interests
 * promotes projects which are more active 

0 stars is normal as we can see, and doesnt really mean anything. It could be a dormant project, a new project, or something that is just rare. 

Projects with a lot of stars are usually not LRL projects, but NLP projects used by regular developers. Having the star badge makes it easy to see these projects are clearly different from LRL but they are often useful too.

If you are looking for a type of technology you will probably look at all projects in that category (ie android apps) but the stars can give you a quick guess of which projects who might welcome you

Having stars on the list also encourages devs to get more stars on their own project, which is good for them.