renamed persona to audience
Audience means
"the readership of a publication"

Persona means
"the aspect of someone's character that is presented to or perceived by others.

Seems rather fishy for something that we want to be simple and easy to
find, like who would be interested in the project.

Persona is also the name of an authentication system by Mozilla
https://www.mozilla.org/en-US/persona/

When I saw the Personas.md file I thought it was something related to
Mozilla persona. We don't need devs getting confused and thinking
that doc is a persona config  or something.