changed linguist to RAs
in my experience, most PIs delegate the task of finding tools to their students, which is why we have a plethora of PHP/MySQL student projects to build online websites for projects. This means the audience isnt linguists in general, because most of them would complain that our page is too techy, but rather their RAs or lab coordinators who have been tasked with the job of reading technical literature and figuring it out.