Add mention of Persephone
Persephone is designed for situations where training data is limited, perhaps as little as an hour of transcribed speech. Such limitations on data are common in the documentation of low-resource languages. It is possible to use such small amounts of data to train a transcription model that can help aid transcription, yet such technology has not been widely adopted. The goal of Persephone is thus to make state-of-the-art phonemic transcription accessible to people involved in language documentation. It is more flexible than CMU-Sphinx in that it handles a wider range of phenomena (including linguistic tone) and yields good results, as reported in recent work: Adams, Oliver, Trevor Cohn, Graham Neubig, Hilaria Cruz, Steven Bird & Alexis Michaud. 2018. Evaluating phonemic transcription of low-resource tonal languages for language documentation. Proceedings of LREC 2018 (Language Resources and Evaluation Conference), 3356–3365. Miyazaki. https://halshs.archives-ouvertes.fr/halshs-01709648.